// packages/config/src/strategy.schema.d.ts
import { z } from 'zod';
import { BOSConfirmationType, MTFMode } from '@quant/shared';
export declare const StrategyConfigSchema: z.ZodObject<{
    name: z.ZodDefault<z.ZodString>;
    minScore: z.ZodDefault<z.ZodNumber>;
    requireLiquiditySweep: z.ZodDefault<z.ZodBoolean>;
    requireBOS: z.ZodDefault<z.ZodBoolean>;
    requireFVG: z.ZodDefault<z.ZodBoolean>;
    requireOrderBlock: z.ZodDefault<z.ZodBoolean>;
    requireHTFAlignment: z.ZodDefault<z.ZodBoolean>;
    riskRewardMinimum: z.ZodDefault<z.ZodNumber>;
    riskPerTrade: z.ZodDefault<z.ZodNumber>;
    maxOpenSignals: z.ZodDefault<z.ZodNumber>;
    maxDailyLossPercent: z.ZodDefault<z.ZodNumber>;
    swingLeftBars: z.ZodDefault<z.ZodNumber>;
    swingRightBars: z.ZodDefault<z.ZodNumber>;
    minSwingDistanceAtrMultiplier: z.ZodDefault<z.ZodNumber>;
    bosConfirmationType: z.ZodDefault<z.ZodNativeEnum<typeof BOSConfirmationType>>;
    displacementThresholdAtr: z.ZodDefault<z.ZodNumber>;
    equalHighLowToleranceAtr: z.ZodDefault<z.ZodNumber>;
    fvgMinGapAtr: z.ZodDefault<z.ZodNumber>;
    mtfMode: z.ZodDefault<z.ZodNativeEnum<typeof MTFMode>>;
}, "strip", z.ZodTypeAny, {
    minScore: number;
    name: string;
    displacementThresholdAtr: number;
    equalHighLowToleranceAtr: number;
    requireLiquiditySweep: boolean;
    requireBOS: boolean;
    requireFVG: boolean;
    requireOrderBlock: boolean;
    requireHTFAlignment: boolean;
    riskRewardMinimum: number;
    riskPerTrade: number;
    maxOpenSignals: number;
    maxDailyLossPercent: number;
    swingLeftBars: number;
    swingRightBars: number;
    minSwingDistanceAtrMultiplier: number;
    bosConfirmationType: BOSConfirmationType;
    fvgMinGapAtr: number;
    mtfMode: MTFMode;
}, {
    minScore?: number | undefined;
    name?: string | undefined;
    displacementThresholdAtr?: number | undefined;
    equalHighLowToleranceAtr?: number | undefined;
    requireLiquiditySweep?: boolean | undefined;
    requireBOS?: boolean | undefined;
    requireFVG?: boolean | undefined;
    requireOrderBlock?: boolean | undefined;
    requireHTFAlignment?: boolean | undefined;
    riskRewardMinimum?: number | undefined;
    riskPerTrade?: number | undefined;
    maxOpenSignals?: number | undefined;
    maxDailyLossPercent?: number | undefined;
    swingLeftBars?: number | undefined;
    swingRightBars?: number | undefined;
    minSwingDistanceAtrMultiplier?: number | undefined;
    bosConfirmationType?: BOSConfirmationType | undefined;
    fvgMinGapAtr?: number | undefined;
    mtfMode?: MTFMode | undefined;
}>;
export type StrategyConfig = z.infer<typeof StrategyConfigSchema>;
export declare const DEFAULT_STRATEGY_CONFIG: StrategyConfig;